+++ cmake/DownloadAndExtract.cmake
@@ -11,7 +11,7 @@ macro(download_and_extract _url _checkfile)
     if(NOT EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/${_base}")
       message(STATUS "Downloading ${_url} -> ${_base}")
       file(DOWNLOAD "${_url}" "${CMAKE_CURRENT_SOURCE_DIR}/${_base}"
-           SHOW_PROGRESS STATUS _rtn)
+            SHOW_PROGRESS STATUS _rtn)
       list(GET _rtn 0 _rtncode)
       if(NOT 0 EQUAL _rtncode)
         message(FATAL_ERROR ${_rtn})
@@ -24,15 +24,16 @@ macro(download_and_extract _url _checkfile)
   endif()
 endmacro()
 
-
-macro(download_platform _base_url _base_name _ext_src _ext_plat)
-  # first download the src file
+# download the src file
+macro(download_src _base_url _base_name _ext_src)
   set(_url "${_base_url}/${_base_name}.tar.gz")
   set(_checkfile "${_base_name}${_ext_src}")
   download_and_extract("${_url}" "${_checkfile}")
-  # now download the platform specific file
+endmacro()
+
+macro(download_platform_specific _base_url _base_name _ext_plat)
+  # download the platform specific file
   set(_url "${_base_url}/${_base_name}-${PYNE_ASM_PLATFORM}.tar.gz")
   set(_checkfile "${_base_name}-${PYNE_ASM_PLATFORM}${_ext_plat}")
   download_and_extract("${_url}" "${_checkfile}")
 endmacro()
-