+++ tests/test_dagmc.py
@@ -5,20 +5,16 @@ import pytest
 from numpy.testing import assert_array_equal
 import multiprocessing
 import numpy as np
-
 from pyne.mesh import HAVE_PYMOAB
-
 from pyne.utils import QAWarning
 
 warnings.simplefilter("ignore", QAWarning)
 
-try:
-    from pyne.mesh import Mesh
+from pyne.mesh import Mesh
+from pyne.dagmc import HAVE_DAGMC
 
-    # See if dagmc module exists but do not import it
-    from pyne import dagmc
-except ImportError:
-    raise pytest.skip(allow_module_level=True)
+if not HAVE_DAGMC:
+    pytest.skip("No DAGMC. Skipping tests", allow_module_level=True)
 
 STRING_TYPES = (str,)