+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Next Version
    * Add dockerfile and workflow to build PyNE with Conda (#1540)
    * Support Cython3+ (#1528)
    * Remove deprecated references to ``basestring`` (#1544)
+   * Redesign Build System for PyNE with Scikit-Build (#1549)
 
 **Fix**
    * Fix Type Mismatch Error in PyNE's ENSDF Processing Module (#1519)
@@ -27,6 +28,7 @@ v0.7.8
    * Move tests from nose to pytest (#1478 #1493)
    * Update MOAB dead link and add PyNE logo to readme file (#1481)
    * Install packages with conda in Dockerfile (#1509 #1510 #1511)
+
 
 **Fix**
    * Add missing rxname offsets (#1482)